fix created Module
