+++ package.json
@@ -9,7 +9,7 @@
     "build-start": "next build && next start"
   },
   "dependencies": {
-    "@material-ui/core": "4.10.0",
+    "@material-ui/core": "^4.10.0",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.56",
     "axios": "^0.20.0",
