+++ src/components/base/ConfirmDialog.module.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+export interface ConfirmDialogProps {
+    confirmDialog: ConfirmDialogObj;
+    setConfirmDialog;
+  }
+  
+  export interface ConfirmDialogObj {
+    title: string;
+    subtitle: string;
+    isOpen: boolean;
+    onConfirm?: () => void;
+  }
