+++ src/components/base/ConfirmDialog.tsx
@@ -7,6 +7,7 @@ import {
   makeStyles,
 } from "@material-ui/core";
 import Controls from "./controls/Controls";
+import { ConfirmDialogProps } from "./ConfirmDialog.module";
 
 const useStyles = makeStyles((theme) => ({
   dialog: {
@@ -21,18 +22,6 @@ const useStyles = makeStyles((theme) => ({
     justifyContent: "center",
   },
 }));
-
-interface ConfirmDialogProps {
-  confirmDialog: ConfirmDialogObj;
-  setConfirmDialog;
-}
-
-export interface ConfirmDialogObj {
-  title: string;
-  subtitle: string;
-  isOpen: boolean;
-  onConfirm?: () => void;
-}
 
 export default function ConfirmDialog(props: ConfirmDialogProps) {
   const { confirmDialog, setConfirmDialog } = props;
