+++ src/components/charts/basic-barchart.module.tsx
@@ -1,0 +1,8 @@
+
+export interface BarChartProps {
+    title: string;
+    color: string;
+    flipped?: boolean;
+    xAxis: string[];
+    yAxis: number[];
+  }
