+++ src/components/goods/add-goods.module.tsx
@@ -1,0 +1,14 @@
+export interface metaDef {
+    productNameToAttributesMap: any;
+    unitList: string[];
+    attributeToDataTypeMap: any;
+  }
+  
+ export interface IBasicProductInfo {
+    amount?: string;
+    unit?: string;
+    product?: string;
+    id?: string;
+    pdc?: string;
+  }
+
