+++ src/components/goods/add-goods.tsx
@@ -9,20 +9,8 @@ import MetaInfoService from "../services/metainfo-service";
 import ProductService from "../services/product-service";
 import BasicProductForm from "./basic-product-info-form";
 import DynamicProductInfo from "./dynamic-product-info";
+import { IBasicProductInfo, metaDef } from "./add-goods.module";
 
-interface metaDef {
-  productNameToAttributesMap: any;
-  unitList: string[];
-  attributeToDataTypeMap: any;
-}
-
-interface IBasicProductInfo {
-  amount?: string;
-  unit?: string;
-  product?: string;
-  id?: string;
-  pdc?: string;
-}
 const initialBasicValues = {
   amount: "",
   unit: "",
