+++ src/components/goods/product-inbox.module.tsx
@@ -1,0 +1,8 @@
+export interface InboxItem {
+    actualOwner: string;
+    alarmFlag: boolean;
+    amount: number;
+    key: string;
+    productName: string;
+    unit: string;
+  }
