+++ src/components/goods/product-inbox.tsx
@@ -1,19 +1,13 @@
 import { IconButton, TableBody, TableCell, TableRow } from "@material-ui/core";
 import { Check, Clear } from "@material-ui/icons";
 import React, { useEffect, useState } from "react";
-import ConfirmDialog, { ConfirmDialogObj } from "../base/ConfirmDialog";
+import ConfirmDialog from "../base/ConfirmDialog";
+import { ConfirmDialogObj } from "../base/ConfirmDialog.module";
 import SearchInputField from "../base/searchInput";
 import useTable from "../base/useTable";
 import productService from "../services/product-service";
+import { InboxItem } from "./product-inbox.module";
 
-interface InboxItem {
-  actualOwner: string;
-  alarmFlag: boolean;
-  amount: number;
-  key: string;
-  productName: string;
-  unit: string;
-}
 
 const ProductInbox = () => {
   const [productState, setProductState] = useState<InboxItem[]>([]);
