+++ src/components/goods/product-outbox.module.tsx
@@ -1,0 +1,8 @@
+export interface OutboxItem {
+    receiver: string;
+    alarmFlag: boolean;
+    amount: number;
+    key: string;
+    productName: string;
+    unit: string;
+  }
