+++ src/components/goods/product-outbox.tsx
@@ -1,19 +1,13 @@
 import { IconButton, TableBody, TableCell, TableRow } from "@material-ui/core";
 import { ArrowBack, Delete, ReportProblem } from "@material-ui/icons";
 import React, { useEffect, useState } from "react";
-import ConfirmDialog, { ConfirmDialogObj } from "../base/ConfirmDialog";
+import ConfirmDialog from "../base/ConfirmDialog";
+import { ConfirmDialogObj } from "../base/ConfirmDialog.module";
 import SearchInputField from "../base/searchInput";
 import useTable from "../base/useTable";
 import productService from "../services/product-service";
+import { OutboxItem } from "./product-outbox.module";
 
-interface OutboxItem {
-  receiver: string;
-  alarmFlag: boolean;
-  amount: number;
-  key: string;
-  productName: string;
-  unit: string;
-}
 
 const ProductOutbox = () => {
   const [productState, setProductState] = useState<OutboxItem[]>([]);
