+++ src/components/goods/set-receiver-dialog.module.tsx
@@ -1,0 +1,8 @@
+import { AxiosResponse } from "axios";
+
+export interface SetReceiverDialogProps {
+    open: boolean;
+    handleClose: Function;
+    handleSubmit: (id: string, receiver: string) => Promise<AxiosResponse>;
+    productId: string;
+  }
