+++ src/components/metainfo/add-attribute-def-dialog.module.tsx
@@ -1,0 +1,10 @@
+import { AxiosResponse } from "axios";
+import { AttributeDefinition } from "../../model";
+
+export interface AddAttributeDefProps {
+    open: boolean;
+    handleClose: Function;
+    handleSubmit: (
+      attributeDefinition: AttributeDefinition
+    ) => Promise<AxiosResponse>;
+  }
