+++ src/components/metainfo/add-product-def-dialog.module.tsx
@@ -1,0 +1,6 @@
+export interface AddProductDefProps {
+    open: boolean;
+    handleClose: Function;
+    handleSubmit: Function;
+    attributes: string[];
+  }
