+++ src/components/metainfo/meta-attribute.module.tsx
@@ -1,0 +1,3 @@
+export interface MetaAttributesProps {
+    attributeToDataTypeMap: any;
+  }
