+++ src/components/metainfo/meta-attributes.tsx
@@ -5,10 +5,9 @@ import {
   TableHead,
   TableRow,
 } from "@material-ui/core";
+import { MetaAttributesProps } from "./meta-attribute.module";
 
-export interface MetaAttributesProps {
-  attributeToDataTypeMap: any;
-}
+
 
 const MetaAttributes = ({ attributeToDataTypeMap }: MetaAttributesProps) => {
   return (
