+++ src/components/user/add-whitelist-to-user-dialog.module.tsx
@@ -1,0 +1,10 @@
+import { AxiosResponse } from "axios";
+
+export interface AddWhitelistToUserDialogProps {
+    open: boolean;
+    handleClose: Function;
+    handleSubmit: (any) => Promise<AxiosResponse>;
+    username: string;
+    userWhitelists: string[];
+    allWhitelists: string[];
+  }
