+++ src/components/user/create-new-user.module.tsx
@@ -1,0 +1,9 @@
+export interface CreateUserProps {
+    existingUsernames: string[];
+  }
+
+export interface NewUser {
+    username: string;
+    password: string;
+    confirmPassword: string;
+  }
