+++ src/components/user/users-management.tsx
@@ -11,7 +11,8 @@ import {
 import { Delete, Info } from "@material-ui/icons";
 import Link from "next/link";
 import React, { useEffect, useState } from "react";
-import ConfirmDialog, { ConfirmDialogObj } from "../base/ConfirmDialog";
+import ConfirmDialog from "../base/ConfirmDialog";
+import { ConfirmDialogObj } from "../base/ConfirmDialog.module";
 import Controls from "../base/controls/Controls";
 import userManagementService from "../services/user-management-service";
