+++ src/components/whitelists/whitelist-management.tsx
@@ -12,7 +12,8 @@ import {
 import { Delete, Info } from "@material-ui/icons";
 import Link from "next/link";
 import React, { useEffect, useState } from "react";
-import ConfirmDialog, { ConfirmDialogObj } from "../base/ConfirmDialog";
+import ConfirmDialog from "../base/ConfirmDialog";
+import { ConfirmDialogObj } from "../base/ConfirmDialog.module";
 import userManagementService from "../services/user-management-service";
 import CreateWhitelistDialog from "./create-whitelist-dialog";
